+++ Version_History.md
@@ -7,6 +7,9 @@ _Oct, 2022_
 ## Alpha 0.2 update:
 
 - Working on contour finetuning
+  - Black keys contour finetuning is done.
+  - White keys contour fintuning is in progress
+    - Needs to adjust from the previous case
 - Preliminary contour creation process adjusted.
   - Now it only creates contours, there is no need to attatch it to a midi number yet.
   - Fusion of contour with midi number (as array index) to be made after finetuning.